big cleanup - move sources under background
remove all the downloading stuff that is now being handled centrally
