+++ background/sources/shared.py
